+++ tests/metrics/test_metastatus_lib.py
@@ -917,7 +917,7 @@ def test_reserved_maintenence_resources_ignores_non_maintenance():
 
 def test_suffixed_number_value():
     assert metastatus_lib.suffixed_number_value("5k") == 5 * 1000
-    assert metastatus_lib.suffixed_number_value("5m") == int(5 * 1000 ** -1)
+    assert metastatus_lib.suffixed_number_value("5m") == 5 * 1000 ** -1
     assert metastatus_lib.suffixed_number_value("5M") == 5 * 1000 ** 2
     assert metastatus_lib.suffixed_number_value("5G") == 5 * 1000 ** 3
     assert metastatus_lib.suffixed_number_value("5T") == 5 * 1000 ** 4
